requirements.txt
# Core Dependencies
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.6.2,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Image Processing
pillow>=11.3.0,<12.0.0

# ML/DL
transformers>=4.55.4,<5.0.0
torch>=2.8.0,<3.0.0
sentencepiece>=0.1.99,<1.0.0
protobuf>=4.0.0,<5.0.0

# WebRTC and WebSockets
# aiohttp and websockets are already dependencies of fastapi/uvicorn
# Only include if you use them directly in your code
aiohttp>=3.9.0,<4.0.0
websockets>=12.0,<13.0
